move PsiTreeUtil and PsiElement extension methods out of CAOS package
