+++ src/main/java/com/badahori/creatures/plugins/intellij/agenteering/caos/def/highlighting/CaosDefHighlighterAnnotator.kt
@@ -8,7 +8,7 @@ import com.intellij.lang.annotation.Annotator
 import com.intellij.openapi.editor.colors.TextAttributesKey
 import com.intellij.psi.PsiElement
 import com.badahori.creatures.plugins.intellij.agenteering.caos.def.psi.api.*
-import com.badahori.creatures.plugins.intellij.agenteering.caos.psi.util.elementType
+import com.badahori.creatures.plugins.intellij.agenteering.utils.elementType
 import com.badahori.creatures.plugins.intellij.agenteering.utils.hasParentOfType
 import com.intellij.openapi.editor.DefaultLanguageHighlighterColors
 import com.intellij.psi.util.PsiTreeUtil
