+++ src/main/java/com/badahori/creatures/plugins/intellij/agenteering/caos/fixes/CaosScriptEtchOnC1QuickFix.kt
@@ -23,8 +23,8 @@ import com.intellij.psi.PsiFile
 import com.intellij.psi.SmartPointerManager
 import com.intellij.psi.codeStyle.CodeStyleManager
 import com.intellij.psi.util.PsiTreeUtil
-import com.badahori.creatures.plugins.intellij.agenteering.caos.psi.util.endOffset
-import com.badahori.creatures.plugins.intellij.agenteering.caos.psi.util.startOffset
+import com.badahori.creatures.plugins.intellij.agenteering.utils.endOffset
+import com.badahori.creatures.plugins.intellij.agenteering.utils.startOffset
 
 /**
  * Turns an ETCH..NEXT statement on C1 to ENUM..NEXT with DOIF TOUC
