+++ src/main/java/com/badahori/creatures/plugins/intellij/agenteering/caos/psi/api/CaosScriptHasCodeBlock.kt
@@ -2,7 +2,7 @@ package com.badahori.creatures.plugins.intellij.agenteering.caos.psi.api
 
 import com.badahori.creatures.plugins.intellij.agenteering.caos.deducer.CaosScope
 import com.badahori.creatures.plugins.intellij.agenteering.caos.deducer.CaosScriptBlockType
-import com.badahori.creatures.plugins.intellij.agenteering.caos.psi.util.getParentOfType
+import com.badahori.creatures.plugins.intellij.agenteering.utils.getParentOfType
 
 interface CaosScriptHasCodeBlock : CaosScriptCompositeElement {
     val codeBlock: CaosScriptCodeBlock?
